don't recreate symlinks if something is there

+++ web/scripts/ms_install_extensions.sh
@@ -47,7 +47,9 @@ install_ms_extensions_prod () {
 }
 
 install_ms_extensions_dev () {
-    ln -s /mafiascum/extensions /opt/bitnami/phpbb/ext/mafiascum
+    if [[ ! -e /opt/bitnami/phpbb/ext/mafiascum ]]; then
+        ln -s /mafiascum/extensions /opt/bitnami/phpbb/ext/mafiascum
+    fi
 }
 
 if [[ $MAFIASCUM_ENVIRONMENT == 'development' ]] || [[ $MAFIASCUM_ENVIRONMENT == 'dev' ]] || [[ $MAFIASCUM_ENVIRONMENT == 'local' ]]; then
